Exploding ForgeAnalyzer
