+++ forges/src/main/java/com/github/maracas/forges/ClientFetcher.java
@@ -1,0 +1,7 @@
+package com.github.maracas.forges;
+
+import java.util.List;
+
+public interface ClientFetcher {
+	List<Commit> fetchClients(Repository repository, Package pkg, int maxClients, int minStars);
+}
