+++ forges/src/main/java/com/github/maracas/forges/report/PullRequestReport.java
@@ -1,11 +1,7 @@
 package com.github.maracas.forges.report;
 
-import com.fasterxml.jackson.databind.ObjectMapper;
-import com.fasterxml.jackson.databind.SerializationFeature;
 import com.github.maracas.forges.PullRequest;
 
-import java.io.File;
-import java.io.IOException;
 import java.util.Collections;
 import java.util.List;
 
@@ -20,10 +16,5 @@ public record PullRequestReport(
 
 	public static PullRequestReport success(PullRequest pr, List<PackageReport> packageReports) {
 		return new PullRequestReport(pr, packageReports, null);
-	}
-
-	public void writeJson(File json) throws IOException {
-		ObjectMapper objectMapper = new ObjectMapper();
-		objectMapper.writeValue(json, this);
 	}
 }
